:recycle: Refactor Tile Iterator + Add Dask Compatible Iterator
